+++ CHANGELOG.rst
@@ -2,6 +2,16 @@
 Changelog
 =========
 
+Version 0.2
+===========
+
+- Added a `to_dict()` function
+
+Version 0.1.1
+=============
+
+- Allow for flexible comment character in `comment(...)`
+
 Version 0.1
 ===========